requirements.txt
# Core dependencies for ECG classification project
tensorflow>=2.16.0
numpy>=1.21.0
pandas>=2.0.0
matplotlib>=3.5.0
scikit-learn>=1.3.0
wfdb>=4.3.0
opencv-python>=4.5.0
seaborn>=0.11.0
Pillow>=8.0.0
kaggle>=1.7.0
